Checking for errors in filter-relevant.sh

+++ test/massive/filter-relevant.sh
@@ -14,12 +14,21 @@ find_tag_by_commit() {
 get_package() {
     local URL="$1" TAG="$2"
 
-    local TMP_DIR="$(mktemp -d)"
-    git clone "$URL" "$TMP_DIR" &> /dev/null
-    git -C "$TMP_DIR" fetch --tags > /dev/null
-    git -C "$TMP_DIR" reset --hard "$TAG" > /dev/null
+    local TMP_DIR
+    TMP_DIR="$(mktemp -d)"
+    [ -z $TMP_DIR ] && { echo >&2 Could not create temporary directory - $URL - $TAG; exit 1; }
 
-    git rev-parse --short HEAD | xargs echo "$TMP_DIR"
+    git clone "$URL" "$TMP_DIR" &> /dev/null
+    [ $? -ne 0 ] && { echo >&2 Error while cloning $URL - $TAG; exit 1; }
+
+    git -C "$TMP_DIR" fetch --tags > /dev/null
+    [ $? -ne 0 ] && { echo >&2 Error while fetching tags $URL - $TAG; exit 1; }
+
+    git -C "$TMP_DIR" reset --hard "$TAG" > /dev/null
+    [ $? -ne 0 ] && { echo >&2 Error while resetting $URL - $TAG; exit 1; }
+
+    git -C "$TMP_DIR" rev-parse --short "$TAG" | xargs echo "$TMP_DIR"
+    [ $? -ne 0 ] && { echo >&2 Error while printing hash for $URL - $TAG; exit 1; }
 }
 
 grep_client_api() {
@@ -75,8 +84,15 @@ RESULTS_DIR="results/filter"
 mkdir -p "$RESULTS_DIR"
 
 while read PACKAGE URL TAG; do
+    [ -z $PACKAGE ] && { echo >&2 PACKAGE is empty - $URL - $TAG; exit 1; }
+    [ -z $URL ] && { echo >&2 URL is empty - $PACKAGE - $TAG; exit 1; }
+    [ -z $TAG ] && { echo >&2 TAG is empty - $PACKAGE - $URL; exit 1; }
+
     echo "Cloning $PACKAGE"
     read PACKAGE_ROOT HASH < <(get_package "$URL" "$TAG")
+
+    [ -z $PACKAGE_ROOT ] && { echo >&2 PACKAGE_ROOT is empty - $HASH; exit 1; }
+    [ -z $HASH ] && { echo >&2 HASH is empty - $PACKAGE_ROOT; exit 1; }
 
     echo "Analyzing $PACKAGE"
     cat > transform-filter.jq <<FILTEREOF
@@ -91,8 +107,10 @@ while read PACKAGE URL TAG; do
 FILTEREOF
     grep_client_api "$PACKAGE_ROOT" "$PACKAGE" \
         | jq -f transform-filter.jq > "$RESULTS_DIR/${PACKAGE//\//--}.json"
+    [ $? -ne 0 ] && exit
 
     rm -rf "$PACKAGE_ROOT"
+	unset PACKAGE_ROOT HASH PACKAGE URL TAG
 done
 
 rm transform-filter.jq
